Get debug mode from the HEXCHAT_DEBUG environment variable
Debug mode gets enabled when this variable is set to a nonempty string.

Additionally, if it set to a valid debug domain, that domain will be
used as a set value to G_MESSAGES_DEBUG instead of "all".

Currently recognized domains are "error", "warning", "critical",
"message", "info", "debug", and the special domain "all".